WE-218 메시지 커스텀 훅 생성
메시지 전송할 때 필요한 로직이 중복되어 사용되어 따로 커스텀 훅으로 추출합니다.
